refactor(scaffolding): change directory structure to match best practices
