+++ internal/databox/static/scaffoldings/golang/cmd/root.go
@@ -15,7 +15,7 @@
 // limitations under the License.
 //
 
-package cdk
+package cmd
 
 import (
 	"fmt"
@@ -25,8 +25,8 @@ import (
 	"github.com/spf13/viper"
 )
 
-// RootCmd represents the base command when called without any subcommands
-var RootCmd = &cobra.Command{
+// rootCmd represents the base command when called without any subcommands
+var rootCmd = &cobra.Command{
 	Use:   "[[.Component]]",
 	Short: "A brief description of your component",
 	Long: `A longer description that spans multiple lines and likely contains
@@ -48,34 +48,34 @@ To quickly get going with this new component, start modifying the placeholder co
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the RootCmd.
 func Execute() {
-	err := RootCmd.Execute()
+	err := rootCmd.Execute()
 	if err != nil {
 		os.Exit(1)
 	}
 }
 
 func init() {
-	RootCmd.PersistentFlags().Bool("debug", false,
+	rootCmd.PersistentFlags().Bool("debug", false,
 		"turn on debug logging",
 	)
-	RootCmd.PersistentFlags().Bool("nocolor", false,
+	rootCmd.PersistentFlags().Bool("nocolor", false,
 		"turn off colors",
 	)
-	RootCmd.PersistentFlags().Bool("nocache", false,
+	rootCmd.PersistentFlags().Bool("nocache", false,
 		"turn off caching",
 	)
-	RootCmd.PersistentFlags().Bool("noninteractive", false,
+	rootCmd.PersistentFlags().Bool("noninteractive", false,
 		"turn off interactive mode (disable spinners, prompts, etc.)",
 	)
-	RootCmd.PersistentFlags().Bool("json", false,
+	rootCmd.PersistentFlags().Bool("json", false,
 		"switch commands output from human-readable to json format",
 	)
 
-	errcheckWARN(viper.BindPFlag("debug", RootCmd.PersistentFlags().Lookup("debug")))
-	errcheckWARN(viper.BindPFlag("nocolor", RootCmd.PersistentFlags().Lookup("nocolor")))
-	errcheckWARN(viper.BindPFlag("nocache", RootCmd.PersistentFlags().Lookup("nocache")))
-	errcheckWARN(viper.BindPFlag("noninteractive", RootCmd.PersistentFlags().Lookup("noninteractive")))
-	errcheckWARN(viper.BindPFlag("json", RootCmd.PersistentFlags().Lookup("json")))
+	errcheckWARN(viper.BindPFlag("debug", rootCmd.PersistentFlags().Lookup("debug")))
+	errcheckWARN(viper.BindPFlag("nocolor", rootCmd.PersistentFlags().Lookup("nocolor")))
+	errcheckWARN(viper.BindPFlag("nocache", rootCmd.PersistentFlags().Lookup("nocache")))
+	errcheckWARN(viper.BindPFlag("noninteractive", rootCmd.PersistentFlags().Lookup("noninteractive")))
+	errcheckWARN(viper.BindPFlag("json", rootCmd.PersistentFlags().Lookup("json")))
 
 	// Here you will define your flags and configuration settings.
 }
